@@ -1,13 +1,15 @@
 from pathlib import Path
 
 import pytest
-from helpers.utils import CWD, IgnoreGitConfig, SandboxedGitRepo
 from plumbum import local
+
+from tests.utils import CWD, IgnoreGitConfig, SandboxedGitRepo
 
 
 @pytest.fixture()
 def root_directory() -> Path:
     """Return the root directory of the project."""
+
     return Path(__file__).parent.parent.resolve()
 
 
@@ -33,12 +35,12 @@ def cloned_template_directory(
 ) -> Path:
     """Return a temporary directory with a cloned copy of the repository."""
 
-    prefix = "cloned-template-repo-"
+    tmp_path = tmp_path_factory.mktemp("cloned-template-repo-")
 
-    with tmp_path_factory.mktemp(prefix) as tmp_path:
-        with CWD(root_directory):
-            local.cmd.cp("--parents", *tracked_files, tmp_path)
-        with SandboxedGitRepo(tmp_path):
-            local.cmd.git("add", ".")
-            local.cmd.git("commit", "--message", "Initial commit")
-            yield tmp_path
+    with CWD(root_directory):
+        local.cmd.cp("--parents", *tracked_files, tmp_path)
+
+    with SandboxedGitRepo(tmp_path):
+        local.cmd.git("add", "./")
+        local.cmd.git("commit", "--message", "Initial commit")
+        yield tmp_path
