@@ -2,8 +2,9 @@ from pathlib import Path
 
 import copier
 import pytest
-from helpers.utils import CWD, SandboxedGitRepo
 from plumbum import local
+
+from tests.utils import CWD, SandboxedGitRepo
 
 
 @pytest.fixture()
@@ -11,12 +12,14 @@ def data() -> dict[str, str]:
     """Return a dictionary with the data to be used in the template."""
 
     return {
-        "username": "quickplates",
-        "projectname": "generic-example",
-        "description": "Generic project example 👤",
-        "releases": "true",
+        "accountname": "foo",
+        "projectname": "foo",
+        "description": "Example project",
+        "reponame": "foo",
+        "repourl": "https://github.com/foo/foo",
         "docs": "true",
-        "docsurl": "https://quickplates.github.io/generic-example",
+        "docsurl": "https://foo.github.io/foo",
+        "releases": "true",
     }
 
 
@@ -28,21 +31,20 @@ def copied_template_directory(
 ) -> Path:
     """Return a temporary directory with a copied template."""
 
-    prefix = "copied-template-"
+    tmp_path = tmp_path_factory.mktemp("copied-template-")
 
-    with tmp_path_factory.mktemp(prefix) as tmp_path:
-        copier.run_copy(
-            str(cloned_template_directory),
-            str(tmp_path),
-            data=data,
-            vcs_ref="HEAD",
-            quiet=True,
-        )
+    copier.run_copy(
+        str(cloned_template_directory),
+        str(tmp_path),
+        data=data,
+        vcs_ref="HEAD",
+        quiet=True,
+    )
 
-        with SandboxedGitRepo(tmp_path):
-            local.cmd.git("add", ".")
-            local.cmd.git("commit", "--message", "Initial commit")
-            yield tmp_path
+    with SandboxedGitRepo(tmp_path):
+        local.cmd.git("add", "./")
+        local.cmd.git("commit", "--message", "Initial commit")
+        yield tmp_path
 
 
 def test_lint(copied_template_directory: Path) -> None:
@@ -51,7 +53,7 @@ def test_lint(copied_template_directory: Path) -> None:
     with CWD(copied_template_directory):
         local.cmd.nix(
             "develop",
-            ".#lint",
+            "./#lint",
             "--command",
             "--",
             "task",
